model drop down menu is functional
Added fetchModels() function to api.py Added populateModelDropdown() to videoHandler.js
