+++ static/index.html
@@ -35,9 +35,7 @@
             <h3>Model Inputs</h3>
             <div class="model-input">Model File Input</div>
             <select id="modelDropdown">
-              <option value="">Select Model</option>
-              <option value="placeholder_model1.pt">Placeholder Model 1</option>
-              <option value="placeholder_model2.pt">Placeholder Model 2</option>
+              <label for="modelDropdown">Select model:</label>
             </select>
           </div>
         </div>
